+++ packages/terra-action-header/CHANGELOG.md
@@ -4,7 +4,7 @@ ChangeLog
 Unreleased
 ----------
 ### Changed
-* Disable the next-previous button group button if associating prop is not set.
+* Disable the next-previous button group button if its associating prop is not set.
 
 1.15.0 - (July 25, 2018)
 ------------------